introduces lmcs_16bpc,  lmcs_16bpc_8pix, lmcs_8bpc
Creates three functions lmcs_16bpc (for width 16 to 128) & lmcs_16bpc_8pix (for width 8) and lmcs_8bpc (widths 32 to 128) in x86 asm to perform LMCS.

The key instruction is vpgatherdd to perform a gather based on indices. Since vpgatherdd only operates on 32bit ints, the major steps are to unpack the 16bit data into 32bits (via punpcklwd with a vec of 0's), perform the gather, shuffle off garbage data, and repack to 16bit ints.

For 16 bits:
Width 16 through 128 will use YMM register (16 * 16 = 256). For 8 pixels, we load 2 rows into a YMM register (for a total of 4 rows). Currently the plan for width 4 is to just downshift to the scalar version.

Rough draft is functional but will require cleanup and further optimizations.