vvc_deblock.asm: removed one-sided IFDEF for vertical
SPATIAL_ACTIVITY (and CHROMA_DEBLOCK) had an ifdef that removed the one-sided calculation. This block of code as activated in the horizontal calculation and not the vertical, hence the one-sided failure.

Since we don't need to ifdef this anymore (this was done for development purposes when implementing vertical), remove the extraneous parameter.